app/requirements.txt
fastapi==0.111.1
uvicorn==0.30.3
gunicorn==22.0.0
pydantic_settings==2.1.0
pydantic[email]==2.5.2
tortoise-orm==0.21.5
aerich==0.7.2
asyncpg==0.29.0

python-dotenv==1.0.1

pytest-asyncio==0.23.8
pytest==8.3.2
Faker==0.7.4